+++ MayhemBTH2017/MayhemBTH2017/ParticleSystem.h
@@ -61,7 +61,7 @@ private:
 	GLuint m_velocityID;
 	GLuint m_timeAliveID;
 	GLuint m_timerID;
-	uint64_t m_offset = 0;
+	uint32_t m_offset = 0;
 	GLfloat m_timer = 0.0f;
 	Transform tmpTransform;
 	GLuint m_drawVAO;
@@ -70,7 +70,7 @@ private:
 	GLfloat info[6];
 	GLuint particlePosID, particleDirID, particleColorID, particleVelocityID, particleTimeAliveID;
 	GLuint inPosID, inDirID, inColID, inTimeID, inSizeID;
-	uint64_t offset;
+	uint32_t offset;
 
 };
